pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
